Rename files, move to two-page fragment. add recents view.
