+++ Chandler/Chandler.csproj
@@ -66,9 +66,10 @@
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MainActivity.cs" />
-    <Compile Include="TimeEntriesPagerAdapter.cs" />
-    <Compile Include="TimeEntryFragment.cs" />
     <Compile Include="Common.cs" />
+    <Compile Include="PagesAdapter.cs" />
+    <Compile Include="TimerFragment.cs" />
+    <Compile Include="ListFragment.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
@@ -99,6 +100,7 @@
     <AndroidResource Include="Resources\drawable\roundedbutton.xml" />
     <AndroidResource Include="Resources\values\Dimensions.xml" />
     <AndroidResource Include="Resources\layout\TimeEntryFragment.axml" />
+    <AndroidResource Include="Resources\layout\ListFragment.axml" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Novell\Novell.MonoDroid.CSharp.targets" />
   <ItemGroup>
