+++ Chandler/Resources/layout/Main.axml
@@ -13,24 +13,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:keepScreenOn="true" />
-    <FrameLayout
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:layout_alignParentBottom="true">
-        <android.support.wearable.view.DotsPageIndicator
-            android:id="@+id/indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginBottom="10dp" />
-        <ImageButton
-            android:background="@drawable/roundedbutton"
-            android:id="@+id/testButton"
-            android:layout_gravity="center_horizontal"
-            android:layout_width="@dimen/roundButtonSize"
-            android:layout_height="@dimen/roundButtonSize"
-            android:layout_marginTop="10dp"
-            android:layout_marginBottom="10dp"
-            android:src="@drawable/icplay" />
-    </FrameLayout>
 </RelativeLayout>
